+++ CHANGES.rst
@@ -81,6 +81,8 @@ Unreleased
 -   ``Pagination`` ``max_per_page`` defaults to 100. :issue:`1091`
 -   ``SQLALCHEMY_RECORD_QUERIES`` is not enabled automatically with ``app.debug`` or
     ``app.testing``. :issue:`1092`
+-   MySQL engines don't set a default for ``pool_recycle`` if not using a queue pool.
+    :issue:`803`
 
 
 Version 2.5.1